+++ README.md
@@ -1,9 +1,11 @@
-# aerogear-android-offline [![Build Status](https://travis-ci.org/aerogear/aerogear-android-offline.png)](https://travis-ci.org/aerogear/aerogear-android-offline)
+# aerogear-android-offline
+ 
+[![Build Status](https://travis-ci.org/aerogear/aerogear-android-offline.png)](https://travis-ci.org/aerogear/aerogear-android-offline)
 
 |                 | Project Info  |
 | --------------- | ------------- |
 | License:        | Apache License, Version 2.0  |
-| Build:          | Maven  |
+| Build:          | Gradle  |
 | Documentation:  | http://aerogear.org/docs/guides/aerogear-android/  |
 | Issue tracker:  | https://issues.jboss.org/browse/AGDROID  |
 | Mailing lists:  | [aerogear-users](http://aerogear-users.1116366.n5.nabble.com/) ([subscribe](https://lists.jboss.org/mailman/listinfo/aerogear-users))  |
