--- aerogear-android-offline-test/pom.xml
@@ -1,148 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  JBoss, Home of Professional Open Source
-  Copyright Red Hat, Inc., and individual contributors
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>aerogear-android-offline-test</artifactId>
-    <packaging>apk</packaging>
-    <name>AeroGear Android Offline Test</name>
-    <url>http://aerogear.org</url>
-
-    <parent>
-        <groupId>org.jboss.aerogear</groupId>
-        <artifactId>aerogear-android-offline-parent</artifactId>
-        <version>3.0.0-SNAPSHOT</version>
-    </parent>
-
-    <properties>
-
-        <!-- Dependencies versions -->
-        <mockito.version>1.9.5</mockito.version>
-        <dexmaker.version>1.2</dexmaker.version>
-        <dexmaker.mockito.version>1.2</dexmaker.mockito.version>
-        <android.support.test.version>0.5</android.support.test.version>
-
-        <!-- Plugins parameters -->
-        <android.maven.plugin.dex.xms>-Xms256m</android.maven.plugin.dex.xms>
-        <android.maven.plugin.dex.xmx>-Xmx2048m</android.maven.plugin.dex.xmx>
-
-    </properties>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>org.jboss.aerogear</groupId>
-            <artifactId>aerogear-android-offline</artifactId>
-            <version>${project.version}</version>
-            <type>aar</type>
-        </dependency>
-
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-all</artifactId>
-            <version>${mockito.version}</version>
-            <exclusions>
-                <exclusion>
-                    <artifactId>hamcrest-core</artifactId>
-                    <groupId>org.hamcrest</groupId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.dexmaker</groupId>
-            <artifactId>dexmaker</artifactId>
-            <version>${dexmaker.version}</version>
-            <exclusions>
-                <exclusion>
-                    <artifactId>hamcrest-core</artifactId>
-                    <groupId>org.hamcrest</groupId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.dexmaker</groupId>
-            <artifactId>dexmaker-mockito</artifactId>
-            <version>${dexmaker.mockito.version}</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.mockito</groupId>
-                    <artifactId>mockito-core</artifactId>
-                </exclusion>
-                <exclusion>
-                    <artifactId>hamcrest-core</artifactId>
-                    <groupId>org.hamcrest</groupId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>com.android.support.test</groupId>
-            <artifactId>runner</artifactId>
-            <version>${android.support.test.version}</version>
-            <type>aar</type>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.hamcrest</groupId>
-                    <artifactId>hamcrest-core</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>junit</groupId>
-                    <artifactId>junit-dep</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-
-        <dependency>
-            <groupId>com.android.support.test</groupId>
-            <artifactId>rules</artifactId>
-            <version>${android.support.test.version}</version>
-            <type>aar</type>
-        </dependency>
-
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>com.simpligility.maven.plugins</groupId>
-                <artifactId>android-maven-plugin</artifactId>
-                <version>${maven.android.plugin.version}</version>
-                <configuration>
-                    <test>
-                        <debug>${android.debug}</debug>
-                    </test>
-                    <sdk>
-                        <platform>${android.platform}</platform>
-                    </sdk>
-                    <proguard>
-                        <skip>${proguard.skip}</skip>
-                    </proguard>
-                    <undeployBeforeDeploy>true</undeployBeforeDeploy>
-                    <dex>
-                        <jvmArguments>
-                            <jvmArgument>${android.maven.plugin.dex.xms}</jvmArgument>
-                            <jvmArgument>${android.maven.plugin.dex.xmx}</jvmArgument>
-                        </jvmArguments>
-                    </dex>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
