--- aerogear-android-offline-test/src/main/AndroidManifest.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-
-    JBoss, Home of Professional Open Source
-    Copyright Red Hat, Inc., and individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-    	http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="org.jboss.aerogear.android.offline.test">
-    <uses-sdk android:minSdkVersion="16" android:targetSdkVersion="24"/>
-
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
-    <uses-permission android:name="android.permission.DISABLE_KEYGUARD"/>
-    <uses-permission android:name="android.permission.WAKE_LOCK"/>
-    <instrumentation
-            android:name="android.support.test.runner.AndroidJUnitRunner"
-            android:targetPackage="org.jboss.aerogear.android.offline.test"/>
-
-    <application
-            android:icon="@drawable/ic_launcher"
-            android:label="@string/app_name">
-        <uses-library android:name="android.test.runner"/>
-
-        <activity
-                android:name="org.jboss.aerogear.android.offline.test.MainActivity"
-                android:label="@string/app_name">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LAUNCHER"/>
-            </intent-filter>
-        </activity>
-    </application>
-
-</manifest>
