--- aerogear-android-offline/pom.xml
@@ -1,160 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  JBoss, Home of Professional Open Source
-  Copyright Red Hat, Inc., and individual contributors
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.jboss.aerogear</groupId>
-    <artifactId>aerogear-android-offline</artifactId>
-    <packaging>aar</packaging>
-    <name>AeroGear Android Offline Library</name>
-    <url>http://aerogear.org</url>
-
-    <parent>
-        <groupId>org.jboss.aerogear</groupId>
-        <artifactId>aerogear-android-offline-parent</artifactId>
-        <version>3.0.0-SNAPSHOT</version>
-    </parent>
-
-    <scm>
-        <connection>scm:git:git@github.com:aerogear/aerogear-android-offline.git</connection>
-        <developerConnection>scm:git:git@github.com:aerogear/aerogear-android-offline.git</developerConnection>
-        <url>git@github.com:aerogear/aerogear-android-offline.git</url>
-        <tag>HEAD</tag>
-    </scm>
-
-    <properties>
-
-        <!-- Dependencies versions -->
-        <aerogear.android.core.version>3.1.0</aerogear.android.core.version>
-        <aerogear.android.pipe.version>3.2.0</aerogear.android.pipe.version>
-
-        <!-- Plugins versions -->
-        <maven.javadoc.plugin.version>2.10.4</maven.javadoc.plugin.version>
-        <builder.helper.plugin.version>1.12</builder.helper.plugin.version>
-
-    </properties>
-
-    <dependencies>
-
-        <dependency>
-            <groupId>org.jboss.aerogear</groupId>
-            <artifactId>aerogear-android-core</artifactId>
-            <version>${aerogear.android.core.version}</version>
-            <type>aar</type>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.aerogear</groupId>
-            <artifactId>aerogear-android-core</artifactId>
-            <version>${aerogear.android.core.version}</version>
-            <type>jar</type>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.aerogear</groupId>
-            <artifactId>aerogear-android-pipe</artifactId>
-            <version>${aerogear.android.pipe.version}</version>
-            <type>aar</type>
-        </dependency>
-
-        <dependency>
-            <groupId>org.jboss.aerogear</groupId>
-            <artifactId>aerogear-android-pipe</artifactId>
-            <version>${aerogear.android.pipe.version}</version>
-            <scope>provided</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>com.simpligility.maven.plugins</groupId>
-                <artifactId>android-maven-plugin</artifactId>
-                <version>${maven.android.plugin.version}</version>
-                <configuration>
-                    <run>
-                        <debug>${android.debug}</debug>
-                    </run>
-                    <sdk>
-                        <platform>${android.platform}</platform>
-                    </sdk>
-                    <proguard>
-                        <skip>${proguard.skip}</skip>
-                    </proguard>
-                </configuration>
-            </plugin>
-
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>${maven.compiler.source}</source>
-                    <target>${maven.compiler.target}</target>
-                </configuration>
-            </plugin>
-
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-javadoc-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>attach-sources</id>
-                        <goals>
-                            <goal>jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-                <configuration>
-                    <sourcepath>${basedir}/src</sourcepath>
-                    <sourceFileExcludes>
-                        <exclude>
-                            **com/google/**
-                        </exclude>
-                        <exclude>**/R.java</exclude>
-                    </sourceFileExcludes>
-                </configuration>
-            </plugin>
-
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>build-helper-maven-plugin</artifactId>
-                <version>${builder.helper.plugin.version}</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>attach-artifact</goal>
-                        </goals>
-                        <configuration>
-                            <artifacts>
-                                <artifact>
-                                    <type>jar</type>
-                                    <file>${project.build.directory}/${project.build.finalName}.jar</file>
-                                </artifact>
-                            </artifacts>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
-
-
