--- aerogear-android-offline/src/main/AndroidManifest.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-
-    JBoss, Home of Professional Open Source
-    Copyright Red Hat, Inc., and individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-    	http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="org.jboss.aerogear.android.offline">
-    <uses-sdk android:minSdkVersion="16" android:targetSdkVersion="24"/>
-</manifest>
