+++ gradle.properties
@@ -1,0 +1,22 @@
+# Android Stuff
+
+VERSION_CODE=001
+VERSION_NAME=0.0.1
+
+# Bintray Stuff
+
+REPO_NAME=aerogear-android-offline
+
+# Maven Stuff
+
+POM_GROUP_ID=org.jboss.aerogear
+POM_DESCRIPTION=AeroGear Android Offline
+POM_URL=http://aerogear.org
+POM_SCM_URL=https://github.com/aerogear/aerogear-android-offline
+POM_SCM_CONNECTION=scm:git@github.com:aerogear/aerogear-android-offline.git
+POM_SCM_DEV_CONNECTION=scm:git@github.com:aerogear/aerogear-android-offline.git
+POM_LICENCE_NAME=The Apache Software License, Version 2.0
+POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
+POM_LICENCE_DIST=repo
+POM_DEVELOPER_ID=aerogear
+POM_DEVELOPER_NAME=aerogear
