+++ library/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=AeroGear Android Offline
+POM_ARTIFACT_ID=aerogear-android-offline
+POM_PACKAGING=aar
