+++ library/src/main/java/org/jboss/aerogear/android/offline/OnStorageCreatedListener.java
