+++ library/src/main/java/org/jboss/aerogear/android/offline/Storage.java
