+++ library/src/main/java/org/jboss/aerogear/android/offline/StorageConfiguration.java
