+++ library/src/main/java/org/jboss/aerogear/android/offline/StorageManager.java
