+++ library/src/main/java/org/jboss/aerogear/android/offline/internal/InternalStorageConfiguration.java
