+++ library/src/main/res/raw/sample_document
