--- pom.xml
@@ -1,121 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  JBoss, Home of Professional Open Source
-  Copyright Red Hat, Inc., and individual contributors
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.jboss.aerogear</groupId>
-    <artifactId>aerogear-android-offline-parent</artifactId>
-    <version>3.0.0-SNAPSHOT</version>
-    <packaging>pom</packaging>
-    <name>AeroGear Android Offline Parent</name>
-    <url>http://aerogear.org</url>
-
-    <parent>
-        <groupId>org.jboss.aerogear</groupId>
-        <artifactId>aerogear-parent</artifactId>
-        <version>0.2.21</version>
-    </parent>
-
-    <modules>
-        <module>aerogear-android-offline</module>
-    </modules>
-
-    <scm>
-        <connection>scm:git:git@github.com:aerogear/aerogear-android-offline.git</connection>
-        <developerConnection>scm:git:git@github.com:aerogear/aerogear-android-offline.git</developerConnection>
-        <url>git@github.com:aerogear/aerogear-android-offline.git</url>
-        <tag>HEAD</tag>
-    </scm>
-
-    <properties>
-
-        <!-- Dependencies versions -->
-        <aerogear.bom.version>0.2.21</aerogear.bom.version>
-
-        <!-- Plugins versions -->
-        <maven.android.plugin.version>4.4.3</maven.android.plugin.version>
-
-        <!-- Android configs -->
-        <android.platform>24</android.platform>
-        <android.debug>false</android.debug>
-        <proguard.skip>true</proguard.skip>
-
-    </properties>
-
-    <dependencyManagement>
-        <dependencies>
-
-            <!-- AeroGear Dependencies BOM Imports -->
-            <dependency>
-                <groupId>org.jboss.aerogear</groupId>
-                <artifactId>aerogear-bom</artifactId>
-                <version>${aerogear.bom.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-
-            <!-- AeroGear Android Dependencies BOM Imports -->
-            <dependency>
-                <groupId>org.jboss.aerogear</groupId>
-                <artifactId>aerogear-android-bom</artifactId>
-                <version>${aerogear.bom.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-
-        </dependencies>
-    </dependencyManagement>
-   
-    <dependencies>
-
-        <!-- Project dependencies -->
-        <dependency>
-            <groupId>android</groupId>
-            <artifactId>android</artifactId>
-            <scope>provided</scope>
-        </dependency>
-   </dependencies>
-
-    <build>
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <groupId>com.simpligility.maven.plugins</groupId>
-                    <artifactId>android-maven-plugin</artifactId>
-                    <version>${maven.android.plugin.version}</version>
-                    <extensions>true</extensions>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-    </build>
-    <profiles>
-        <profile>
-            <id>android-tests</id>
-            <activation>
-                <property>
-                    <name>!release</name>
-                </property>
-            </activation>
-            <modules>
-                <module>aerogear-android-offline-test</module>
-            </modules>
-        </profile>
-    </profiles>
-
-</project>
-
